moved Pos and Size to geom.go
